pyproject.toml
[tool.poetry]
name = "sinacare_algorithms"
version = "0.1.0"
description = ""
authors = ["mhmb <m.h.motie@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Django = "^3.2.9"
psycopg2-binary = "^2.9.2"
pynguin = "^0.22.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.19.0"
coverage = "^6.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
